refactored the code for better readability
