+++ src/common/constants.js
@@ -66,6 +66,36 @@ export const COMBINED_REGEX = new RegExp(
   "g"
 );
 
+export const URL_VALIDATORS = {
+  youtube: url => {
+    const match = url.match(YOUTUBE_URL_REGEXP);
+
+    return match && `https://www.youtube.com/embed/${match[5]}`;
+  },
+  vimeo: url => {
+    const match = url.match(VIMEO_URL_REGEXP);
+
+    return match && `https://player.vimeo.com/video/${match[4]}?h=${match[5]}`;
+  },
+  loom: url => {
+    const match = url.match(LOOM_URL_REGEXP);
+
+    return (
+      match && `https://www.loom.com/embed/${match[4]}?t=${match[5] || ""}`
+    );
+  },
+  neetoRecord: url => {
+    const match = url.match(NEETO_RECORD_URL_REGEXP);
+
+    return match && url.replace("watch", "embeds");
+  },
+  supademo: url => {
+    const match = url.match(SUPA_DEMO_URL_REGEXP);
+
+    return match && `https://app.supademo.com/embed/${match[5]}?embed_v=2`;
+  },
+};
+
 export const EDITOR_SIZES = {
   MEDIUM: "medium",
   LARGE: "large",
